Add files via upload
These seven files are a distribution of DCU2OFX, a little workaround tool to allow DCU members to get around the lack of OFX file support, at least for now.  It consists of seven files.  The script itself, written in powershell, and a text file describing the script and how to use it; three template files that specify the output format; and sample input and output files.  People who have been following the discussion in the "DCU What to do" Facebook group will know what this is about.